requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
Werkzeug==2.3.7
SQLAlchemy>=2.0.16
psycopg2-binary==2.9.9
openpyxl==3.1.2
python-dateutil==2.8.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0